move lines blueprint from yafowil.common to yafowil.lines
