+++ src/yafowil/loader.py
@@ -11,6 +11,7 @@ def register():  # noqa
     import yafowil.email
     import yafowil.field
     import yafowil.file
+    import yafowil.lines
     import yafowil.number
     import yafowil.password
     import yafowil.persistence
