+++ src/yafowil/tests/__init__.py
@@ -8,6 +8,7 @@ from yafowil import datatypes
 from yafowil import email
 from yafowil import field
 from yafowil import file
+from yafowil import lines
 from yafowil import number
 from yafowil import password
 from yafowil import persistence
@@ -36,6 +37,7 @@ class YafowilTestCase(NodeTestCase):
         reload(email)
         reload(field)
         reload(file)
+        reload(lines)
         reload(number)
         reload(password)
         reload(persistence)
@@ -72,6 +74,7 @@ def test_suite():
     from yafowil.tests import test_email
     from yafowil.tests import test_field
     from yafowil.tests import test_file
+    from yafowil.tests import test_lines
     from yafowil.tests import test_number
     from yafowil.tests import test_password
     from yafowil.tests import test_persistence
@@ -95,6 +98,7 @@ def test_suite():
     suite.addTest(unittest.findTestCases(test_email))
     suite.addTest(unittest.findTestCases(test_field))
     suite.addTest(unittest.findTestCases(test_file))
+    suite.addTest(unittest.findTestCases(test_lines))
     suite.addTest(unittest.findTestCases(test_number))
     suite.addTest(unittest.findTestCases(test_password))
     suite.addTest(unittest.findTestCases(test_persistence))
